+++ outputs.tf
@@ -8,3 +8,7 @@ output "nic_id" {
     description = "The ID of the Network Interface"
 }
 
+output "sg_id" {
+    value = azurerm_network_security_group.tamr-vm-sg.id
+    description = "The ID of the security group"
+}