+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+# Tamr Azure VM module - v0.2.0 - May 27th 2020
+* Created security group for VM egress and ingress
 # Tamr Azure VM module - v0.1.0 - May 8th 2020
 ## Added
 * Tamr Azure VM module